==> Code/6_network_analysis.Rmd <==
---
title: "6_network analysis"
author: "Pedro"
date: "10/8/2021"
output: html_document
editor_options: 
  chunk_output_type: console
---

[MA] If coming from script 5, you should have 'physeq_filtered' load in your environment, if not load it here 
```{r}

#Load 'normalization_listed', a ps list with CSS, rarefied and non-rarefied
normalization_listed <- readRDS("./R output/01_phyloseq_objects/normalization_listed.rds")

#let's make it a single ps
physeq_filtered <- normalization_listed$physeq_filtered

#Split by Plant Species and Root compartment 
ps_list_unormalized <- phyloseq_sep_variable(physeq_filtered, variable = "Sp_Comp")

```

[MA] It takes a long time to run the network analysis, for plots you can skip this part and load nodes and edges data already calculated by Pedro in Chunk 14, line XXX


# load libraries and data
```{r}
source("./Code/Functions/load_libraries.R")
load("./R output/analysis_session.RData") #[MA]: This object doesn't exist

```

##### 6.0 - load libraries & functions
```{r}
##### 0 - load libraries and custo functions####
source("./Code/Functions/Network_custom_functions.R") # this is a large set of custom functions used over this script
```


##### 6.1 - prepare and filter phyloseq objects

```{r}

# this makes one list wth 4 objects
ps_list_unormalized <- phyloseq_sep_variable(physeq_filtered, variable = c("Plant_species", "Sample_type"))

# apply filter function on list of phyloseq objects - account for at least 0.5% of the reads in a sample & be present in at least 25% of the samples
ps_list_network <- lapply(ps_list_unormalized, function(x) filterPhyseq(x, 0.005, 25))

# *************************** Done!
```


##### 6.2 - Calculate the network

takes ~2h for 16 netwoks at 1-5% cut offs in 1/4 to 1/2 of samples (total 94 samples in the complete dataset, ~24 samples per network)

```{r}
# define pulsar parameters. when testing different parameters, reduce the rep.num to 50 to speed up calculations
pargs <- list(rep.num = 100, seed = 10010, ncores = 1, thresh = 0.05)

# run the main function to calculate the network. it can take several minutes to a couple hours
list_spiec_objct <- lapply(ps_list_network, function(x) {
  SpiecEasi::spiec.easi(x,
    method = "glasso", # if you change methods to "mb" you will have to change detials in make_igrap() as commented in that function
    lambda.min.ratio = 1e-2, # the higher the more edges, leading to higher density network. as it goes lower, computational times increases SUBSTANTIALLY. standard is 1e-3 but 1e-2 works well ; pedro's PC crahs when 1e-7
    nlambda = 30, # number of slices between minimum amount of edges to maximum. low lavues dive more sparse networks, high values give denser networks. keep it between 10 and 100
    sel.criterion = "bstars", # selects the lambda and modules the most stable connections
    pulsar.select = TRUE,
    pulsar.params = pargs
  )
}) # iteration of starts, n)


# *************************** Done!
```


##### 6.3 - Make igraph object

igraph is another package for netowrks that will let you acess and manipulate nodes, edges and their proprieties

```{r}

# the  make_igraph() custom function makes a weighted igraph object from the spiec.easi object. we need igraph objects to calculate network metrics
listed_igraph <- mapply(function(x, y) {
  make_igraph(x, y)
},
x = list_spiec_objct,
y = ps_list_network,
SIMPLIFY = FALSE
)

# we should now remove the spiec_easy oject ebcause it is really heavy
rm(list_spiec_objct)
# *************************** Done!
```

##### 6.4 - Check fit to power law, node degree distribution, and edge weight distribution

one of the best indicators for the reliability of your network is wether it fits a power law or not (many OTUs with low degree, few OTUs with high degree) while not all microbial networks will follow a power law, it is very convinient when we are looking for keystone species. we adopt this simplification even knowing it might not be a ground truth.

```{r}


# with this, we fit a power law and make a statistical test to the distribution on a list of networks
power_law_list <- lapply(listed_igraph, function(x) fit_power_law(degree(x)))

# this gives you the P values of the power law tests.
map(power_law_list, 6)

# Checks histogram of degree over lists of netwokrs. it should NOT be normal - expect a few OTUs with highe degree and many otus with low degree.
# These functions  provide titles to the histogram lists, which is actually quick tricky to perofrm in R
Degree_histogram_with_name <- function(list_objc) {
  lapply(names(list_objc), function(x) {
    hist(degree(list_objc[[x]]), main = paste(x), xlab = "Node Degree")
  })
}

# Checks histogram of edge weight over lists of netwokrs. it is expected to be normal
Weight_histogram_with_name <- function(list_objc) {
  lapply(names(list_objc), function(x) {
    hist(E(list_objc[[x]])$weight, main = paste(x), xlab = "Edge Weight")
  })
}

# make the plots
Degree_histogram_with_name(listed_igraph)
Weight_histogram_with_name(listed_igraph)
# *************************** Done!
```

##### 6.5 - Compare against a random network

Our network is useless if we can't tell it apart from a random network with the same number of nodes and edges. it is OK to ahve a few metrics in a few networks that are still similar to a random network, but pay attention and record this when selecting your cut-offs

by changing parameters I improved from ~ 20 metrics similar to random to just 2 metrics similar to random.
```{r}

# The Real_VS_random_networks() custom function will compare the real network wiht 100 random networks, returning metrics that are different form random (average +/- SD) as TRUE

# run the custom function on a a list of networks
list_random_VS_real_1000 <- map(listed_igraph, Real_VS_random_networks)

as.data.frame(map(list_random_VS_real_1000, 2)) # is the real network different from the random network?

# *************************** Done!
```

##### 6.6 - Define network node proprieties
Now that we know our networks differ from random we can check their general proprieties. this can help us compare multiple different network we create, telling apart treatment or sample type effects ~ puttig the results on a PCA migt be a good idea.. also, check if your metrics increase or decrease according gradients you may have in your experiments

At core/paths/shortest_paths.c:188 : Weight vector must be non-negative, got -0.126345. Invalid value
```{r}

# on a list of networks
list_partial_network_metrics <- map(listed_igraph, Generate_RealNetworks_metrics)

# we still have a few more metrics to add in the following chunks

# *************************** Done!
```


###### 6.6.1 - Generate node metrics

```{r}
# The Generate_node_metrics2( custom) function will calculate a set of node metrics you may be interested in exploring

node_metrics <- mapply(function(x, y) {
  Generate_node_metrics2(x, y)
},
x = listed_igraph,
y = ps_list_network,
SIMPLIFY = FALSE
)
```



###### 6.6.2 - Calculate Zi and Pi

```{r}

# The Zi_Pi_list() custom function will  will return a list, with Zi/Pi values for nodes as well as a count of module hubs (Zi>2.5) and connectors (Pi>0.62)  it will take ~40 sec to run on a network with 600 nodes and 1000 edges,

# to a list of networks
ZiPi_listed <- map(listed_igraph, Zi_Pi_list)


# check Zi-Pi scatterplot

lapply(ZiPi_listed, function(x) {
  ggplot(x[[1]], aes(x = Z, y = P)) +
    geom_hline(yintercept = 0.62, linetype = "dashed", color = "red") +
    geom_vline(xintercept = 2.5, linetype = "dashed", color = "red") +
    geom_point()
})



# merges node emtrics with ZiPi
node_metrics <- mapply(function(x, y) {
  merge(x, # the Generate_node_metrics2() output
    y[[1]][, c(1:2, 5:6)], # 4 columns of the ZiPi df of OTU metrics
    by.x = 0, # merge by rownames for object x
    by.y = "names", # merge by the collumn named "names" in object y
    all.x = TRUE
  )
}, # keep rows even if they ahve data missing, in this case ZiPi outside the main component
x = node_metrics,
y = ZiPi_listed,
SIMPLIFY = FALSE
)



# this will take the list with the number of connectors and hubs for each ntwork and add it to the main dataframe with netowr metrics
list_network_metrics <- do.call(
  rbind.data.frame,
  Map(c, map(ZiPi_listed, 2), list_partial_network_metrics)
)

#median module size has no variation, causing a PCA to fail. let's remove it
list_network_metrics<-list_network_metrics[,-25]
  

#calculate a PCA to pplot with autoplot
list_network_metrics_pca<-prcomp(list_network_metrics, scale. = TRUE)

# make a PCA of network metrics
autoplot(list_network_metrics_pca,
  label = TRUE,
  loadings = TRUE,
  loadings.colour = "red",
  loadings.label = TRUE,
  loadings.label.size = 3
) +
  theme_bw()


# save these network metrics externally
write.csv(list_network_metrics, "./R output/06_network_analysis/netwrok_metrics.csv")
```


###### 6.6.3 - Dectect and label keystones

```{r}

# The KeystoneDetector3() custom function will plot keystones on a chart, besides providing a list of keystone species for your netowkr

# run the custom function on a list
keystone_list <- map(node_metrics, KeystoneDetector3)

# add the keystone definitions to the node metric dataframe
complete_node_metrics <- mapply(function(x, y) {
  merge(x, # the Generate_node_metrics2() output
    y[[2]][, c(1, 18)], # keeps only the ASV number and keystone definition
    #  by.x = 0, # merge by rownames for object x
    #  by.y = "names", # merge by the collumn named "names" in object y
    all.x = TRUE
  )
}, # keep rows even if they ahve data missing, in this case ZiPi outside the main component
x = node_metrics,
y = keystone_list,
SIMPLIFY = FALSE
)
```

###### 6.6.4 - add OTU abundances to the node metric dataframe

Here we will add the OTU frequency under each treatment and niche to the node table. Will  be usefull on network visualization as we can regulate the size of nodes according their abundances

```{r}

############################################################################################################
############################################################################################################
############################################################################################################
# I could not automate this step due to total brain failure on a friday afternoon. this code chunk is a nightmare!
############################################################################################################
############################################################################################################

# add total number of rarefied sequences of the OTU
ps_list_rarefied # 4-split, rarefied
ps_list_treat # 4 treaments of the 4-split, rarefied


# total number of seuqneces of this OTU in the 4-slice
total_per_slice <- lapply(ps_list_rarefied, function(x) {
  rowSums(otu_table(x))
})


# add this total number of sequences to the node_metrics list
complete_node_metrics <- mapply(function(x, y) {
  merge(x, # the Generate_node_metrics2() output
    y, # keeps only the ASV number and keystone definition
    by.x = "Row.names", # merge by rownames for object x)
    by.y = 0,
    all.x = FALSE
  )
}, # keep rows even if they ahve data missing, in this case ZiPi outside the main component
x = complete_node_metrics,
y = total_per_slice,
SIMPLIFY = FALSE
)

# boooo bad conding to change names of clumns on a list
colnames(complete_node_metrics$Arabidopsis_thaliana.Root)[19] <- "freq_in_4slice"
colnames(complete_node_metrics$Arabidopsis_thaliana.Soil)[19] <- "freq_in_4slice"
colnames(complete_node_metrics$Brassica_oleraceae.Root)[19] <- "freq_in_4slice"
colnames(complete_node_metrics$Brassica_oleraceae.Soil)[19] <- "freq_in_4slice"

# add treatment row sums to AT_root
total_per_treat_AT_root <- lapply(ps_list_treat$Arabidopsis_thaliana.Root, function(x) {
  rowSums(otu_table(x))
})


complete_node_metrics$Arabidopsis_thaliana.Root <- merge(
  x = complete_node_metrics$Arabidopsis_thaliana.Root, # the Generate_node_metrics2() output
  y = total_per_treat_AT_root$control, # keeps only the ASV number and keystone definition
  by.x = "Row.names", # merge by rownames for object x)
  by.y = 0,
  all.x = TRUE
) # keep rows even if
colnames(complete_node_metrics$Arabidopsis_thaliana.Root)[20] <- "freq_in_control"


complete_node_metrics$Arabidopsis_thaliana.Root <- merge(
  x = complete_node_metrics$Arabidopsis_thaliana.Root, # the Generate_node_metrics2() output
  y = total_per_treat_AT_root$MeJA_0.1, # keeps only the ASV number and keystone definition
  by.x = "Row.names", # merge by rownames for object x)
  by.y = 0,
  all.x = TRUE
) # keep rows even if
colnames(complete_node_metrics$Arabidopsis_thaliana.Root)[21] <- "freq_in_MeJA_0.1"

complete_node_metrics$Arabidopsis_thaliana.Root <- merge(
  x = complete_node_metrics$Arabidopsis_thaliana.Root, # the Generate_node_metrics2() output
  y = total_per_treat_AT_root$MeJA_1.0, # keeps only the ASV number and keystone definition
  by.x = "Row.names", # merge by rownames for object x)
  by.y = 0,
  all.x = TRUE
) # keep rows even if
colnames(complete_node_metrics$Arabidopsis_thaliana.Root)[22] <- "freq_in_MeJA_1.0"

complete_node_metrics$Arabidopsis_thaliana.Root <- merge(
  x = complete_node_metrics$Arabidopsis_thaliana.Root, # the Generate_node_metrics2() output
  y = total_per_treat_AT_root$oral_secretion, # keeps only the ASV number and keystone definition
  by.x = "Row.names", # merge by rownames for object x)
  by.y = 0,
  all.x = TRUE
) # keep rows even if
colnames(complete_node_metrics$Arabidopsis_thaliana.Root)[23] <- "freq_in_oral_secretion"



# add treatment row sums to AT_soil
total_per_treat_AT_Soil <- lapply(ps_list_treat$Arabidopsis_thaliana.Soil, function(x) {
  rowSums(otu_table(x))
})


complete_node_metrics$Arabidopsis_thaliana.Soil <- merge(
  x = complete_node_metrics$Arabidopsis_thaliana.Soil, # the Generate_node_metrics2() output
  y = total_per_treat_AT_Soil$control, # keeps only the ASV number and keystone definition
  by.x = "Row.names", # merge by rownames for object x)
  by.y = 0,
  all.x = TRUE
) # keep rows even if
colnames(complete_node_metrics$Arabidopsis_thaliana.Soil)[20] <- "freq_in_control"


complete_node_metrics$Arabidopsis_thaliana.Soil <- merge(
  x = complete_node_metrics$Arabidopsis_thaliana.Soil, # the Generate_node_metrics2() output
  y = total_per_treat_AT_Soil$MeJA_0.1, # keeps only the ASV number and keystone definition
  by.x = "Row.names", # merge by rownames for object x)
  by.y = 0,
  all.x = TRUE
) # keep rows even if
colnames(complete_node_metrics$Arabidopsis_thaliana.Soil)[21] <- "freq_in_MeJA_0.1"

complete_node_metrics$Arabidopsis_thaliana.Soil <- merge(
  x = complete_node_metrics$Arabidopsis_thaliana.Soil, # the Generate_node_metrics2() output
  y = total_per_treat_AT_Soil$MeJA_1.0, # keeps only the ASV number and keystone definition
  by.x = "Row.names", # merge by rownames for object x)
  by.y = 0,
  all.x = TRUE
) # keep rows even if
colnames(complete_node_metrics$Arabidopsis_thaliana.Soil)[22] <- "freq_in_MeJA_1.0"

complete_node_metrics$Arabidopsis_thaliana.Soil <- merge(
  x = complete_node_metrics$Arabidopsis_thaliana.Soil, # the Generate_node_metrics2() output
  y = total_per_treat_AT_Soil$oral_secretion, # keeps only the ASV number and keystone definition
  by.x = "Row.names", # merge by rownames for object x)
  by.y = 0,
  all.x = TRUE
) # keep rows even if
colnames(complete_node_metrics$Arabidopsis_thaliana.Soil)[23] <- "freq_in_oral_secretion"

# add treatment row sums to BO_root
total_per_treBO_BO_root <- lapply(ps_list_treat$Brassica_oleraceae.Root, function(x) {
  rowSums(otu_table(x))
})


complete_node_metrics$Brassica_oleraceae.Root <- merge(
  x = complete_node_metrics$Brassica_oleraceae.Root, # the Generate_node_metrics2() output
  y = total_per_treBO_BO_root$control, # keeps only the ASV number and keystone definition
  by.x = "Row.names", # merge by rownames for object x)
  by.y = 0,
  all.x = TRUE
) # keep rows even if
colnames(complete_node_metrics$Brassica_oleraceae.Root)[20] <- "freq_in_control"


complete_node_metrics$Brassica_oleraceae.Root <- merge(
  x = complete_node_metrics$Brassica_oleraceae.Root, # the Generate_node_metrics2() output
  y = total_per_treBO_BO_root$MeJA_0.1, # keeps only the ASV number and keystone definition
  by.x = "Row.names", # merge by rownames for object x)
  by.y = 0,
  all.x = TRUE
) # keep rows even if
colnames(complete_node_metrics$Brassica_oleraceae.Root)[21] <- "freq_in_MeJA_0.1"

complete_node_metrics$Brassica_oleraceae.Root <- merge(
  x = complete_node_metrics$Brassica_oleraceae.Root, # the Generate_node_metrics2() output
  y = total_per_treBO_BO_root$MeJA_1.0, # keeps only the ASV number and keystone definition
  by.x = "Row.names", # merge by rownames for object x)
  by.y = 0,
  all.x = TRUE
) # keep rows even if
colnames(complete_node_metrics$Brassica_oleraceae.Root)[22] <- "freq_in_MeJA_1.0"

complete_node_metrics$Brassica_oleraceae.Root <- merge(
  x = complete_node_metrics$Brassica_oleraceae.Root, # the Generate_node_metrics2() output
  y = total_per_treBO_BO_root$oral_secretion, # keeps only the ASV number and keystone definition
  by.x = "Row.names", # merge by rownames for object x)
  by.y = 0,
  all.x = TRUE
) # keep rows even if
colnames(complete_node_metrics$Brassica_oleraceae.Root)[23] <- "freq_in_oral_secretion"



# add treatment row sums to BO_soil
total_per_treBO_BO_Soil <- lapply(ps_list_treat$Brassica_oleraceae.Soil, function(x) {
  rowSums(otu_table(x))
})


complete_node_metrics$Brassica_oleraceae.Soil <- merge(
  x = complete_node_metrics$Brassica_oleraceae.Soil, # the Generate_node_metrics2() output
  y = total_per_treBO_BO_Soil$control, # keeps only the ASV number and keystone definition
  by.x = "Row.names", # merge by rownames for object x)
  by.y = 0,
  all.x = TRUE
) # keep rows even if
colnames(complete_node_metrics$Brassica_oleraceae.Soil)[20] <- "freq_in_control"


complete_node_metrics$Brassica_oleraceae.Soil <- merge(
  x = complete_node_metrics$Brassica_oleraceae.Soil, # the Generate_node_metrics2() output
  y = total_per_treBO_BO_Soil$MeJA_0.1, # keeps only the ASV number and keystone definition
  by.x = "Row.names", # merge by rownames for object x)
  by.y = 0,
  all.x = TRUE
) # keep rows even if
colnames(complete_node_metrics$Brassica_oleraceae.Soil)[21] <- "freq_in_MeJA_0.1"

complete_node_metrics$Brassica_oleraceae.Soil <- merge(
  x = complete_node_metrics$Brassica_oleraceae.Soil, # the Generate_node_metrics2() output
  y = total_per_treBO_BO_Soil$MeJA_1.0, # keeps only the ASV number and keystone definition
  by.x = "Row.names", # merge by rownames for object x)
  by.y = 0,
  all.x = TRUE
) # keep rows even if
colnames(complete_node_metrics$Brassica_oleraceae.Soil)[22] <- "freq_in_MeJA_1.0"

complete_node_metrics$Brassica_oleraceae.Soil <- merge(
  x = complete_node_metrics$Brassica_oleraceae.Soil, # the Generate_node_metrics2() output
  y = total_per_treBO_BO_Soil$oral_secretion, # keeps only the ASV number and keystone definition
  by.x = "Row.names", # merge by rownames for object x)
  by.y = 0,
  all.x = TRUE
) # keep rows even if
colnames(complete_node_metrics$Brassica_oleraceae.Soil)[23] <- "freq_in_oral_secretion"



complete_node_metrics

###############################################
###############################################
# nightmare ends here but will haunt us later##
###############################################
###############################################
```


`


##### 6.8 - Export network data
we will now use the edge file to import it as a network file to cytoscape. then we can importe the node proprieties to the enw visualizations
```{r}

# remove NAs as they may crash cytoscape

complete_node_metrics$Arabidopsis_thaliana.Root[is.na(complete_node_metrics$Arabidopsis_thaliana.Root)] <- ""
complete_node_metrics$Arabidopsis_thaliana.Soil[is.na(complete_node_metrics$Arabidopsis_thaliana.Soil)] <- ""
complete_node_metrics$Brassica_oleraceae.Root[is.na(complete_node_metrics$Brassica_oleraceae.Root)] <- ""
complete_node_metrics$Brassica_oleraceae.Soil[is.na(complete_node_metrics$Brassica_oleraceae.Soil)] <- ""

# save these network metrics externally. we will do this manually for only 4 networks
write.csv(
  x = complete_node_metrics$Arabidopsis_thaliana.Root,
  file = "./R output/06_network_analysis/Nodes_AT_root.csv",
  quote = FALSE,
  row.names = FALSE
)
write.csv(
  x = igraph::as_data_frame(listed_igraph$Arabidopsis_thaliana.Root),
  file = "./R output/06_network_analysis/Edges_AT_root.csv",
  quote = FALSE
)


write.csv(
  x = complete_node_metrics$Arabidopsis_thaliana.Soil,
  file = "./R output/06_network_analysis/Nodes_AT_soil.csv",
  quote = FALSE,
  row.names = FALSE
)
write.csv(
  x = igraph::as_data_frame(listed_igraph$Arabidopsis_thaliana.Soil),
  file = "./R output/06_network_analysis/Edges_AT_soil.csv",
  quote = FALSE
)


write.csv(
  x = complete_node_metrics$Brassica_oleraceae.Root,
  file = "./R output/06_network_analysis/Nodes_BO_root.csv",
  quote = FALSE,
  row.names = FALSE
)
write.csv(
  x = igraph::as_data_frame(listed_igraph$Brassica_oleraceae.Root),
  file = "./R output/06_network_analysis/Edges_BO_root.csv",
  quote = FALSE
)


write.csv(
  x = complete_node_metrics$Brassica_oleraceae.Soil,
  file = "./R output/06_network_analysis/Nodes_BO_soil.csv",
  quote = FALSE,
  row.names = FALSE
)
write.csv(
  x = igraph::as_data_frame(listed_igraph$Brassica_oleraceae.Soil),
  file = "./R output/06_network_analysis/Edges_BO_soil.csv",
  quote = FALSE
)
```

#Reading results
[MA] okay, so start here if you'd like to skip calculating the network analysis and instead just look at the plots
Load first the libraries needed 

## Load libraries
```{r}

library(readr) #to open csv's

```

##Nodes
```{r}

#[MA] Individual results
Nodes_AT_root <- read_csv("./R output/06_network_analysis/Nodes_AT_root.csv")
Nodes_BO_root <- read_csv("./R output/06_network_analysis/Nodes_BO_root.csv")
Nodes_AT_soil <- read_csv("./R output/06_network_analysis/Nodes_AT_soil.csv")
Nodes_BO_soil <- read_csv("./R output/06_network_analysis/Nodes_BO_soil.csv")

#[MA] network metrics (Summary)
network_metrics <- read_csv("./R output/06_network_analysis/netwrok_metrics.csv")

#[MA] check colnames
colnames(Nodes_AT_root)
colnames(network_metrics)

#[MA] make a list 
Nodes_list <- list(Nodes_AT_root = Nodes_AT_root,
                   Nodes_BO_root = Nodes_BO_root, 
                   Nodes_AT_soil = Nodes_AT_soil,
                   Nodes_BO_soil = Nodes_BO_soil)

#[MA] check if keystone
keystone_counts <- sapply(Nodes_list, function(df) {
    sum(df$keystone_taxa == "Keystone", na.rm = TRUE)})

Nodes_BO_soil$keystone_taxa
Nodes_BO_soil$keystone_taxa


```


```{r}
#[MA] Now, let's add the classification as module hub or module connector 

#[MA] Function to classify hubs and connectors
add_module_roles <- function(df) {
  df <- df %>%
    mutate(
      module_hub = ifelse(Zi > 2.5, "Module Hub", "Non-Hub"),
      module_connector = ifelse(Pi > 0.62, "Module Connector", "Non-Connector")
    )
  return(df)}

#[MA] Apply to all dataframes in the list
Nodes_list <- lapply(Nodes_list, add_module_roles)

 #Count hubs and connectors per sub-community
module_summary <- lapply(Nodes_list, function(df) {
  data.frame(
    Module_Hubs = sum(df$module_hub == "Module Hub", na.rm = TRUE),
    Module_Connectors = sum(df$module_connector == "Module Connector", na.rm = TRUE)
  )
}) #[MA] ok, perfect, now it matches the `network metrics` summary


#[MA] get info of only those classified as central ASVs (hubs+connectors+keystone)
central_keystone_asv_list <- lapply(Nodes_list, function(df) {
  df %>%
    filter(
      module_hub == "Module Hub" |
      module_connector == "Module Connector" |
      keystone_taxa == "Keystone"
    )
}) #[MA]this is the one Pedro used for the Fisher test

#[MA]make it a single df 
central_keystone_asv_df <- bind_rows(central_keystone_asv_list, .id = "Sub_community")
central_keystone_asv_df  <- as.data.frame(central_keystone_asv_df)
head(central_keystone_asv_df)

#[MA]save this output;
saveRDS(central_keystone_asv_df, "./R output/Network files/ASVs_central_Network_Analysis_df.rds")


```


##Edges
```{r}

#[MA]Individual results
Edges_AT_root <- read_csv("./R output/06_network_analysis/Edges_AT_root.csv")
Edges_BO_root <- read_csv("./R output/06_network_analysis/Edges_BO_root.csv")
Edges_AT_soil <- read_csv("./R output/06_network_analysis/Edges_AT_soil.csv")
Edges_BO_soil <- read_csv("./R output/06_network_analysis/Edges_BO_soil.csv")

#[MA]quick plot
ggplot(Edges_AT_root, aes(x = weight, fill = positive_negative)) +
    geom_histogram(bins = 30, color = "black") +
    labs(x = "Edge weight", y = "Count") +
    theme_minimal()

#[MA]quick plot
Edges_AT_root %>%
    ggplot(aes(x = weighted_edge_betweenness)) +
    geom_histogram(bins = 30, fill = "steelblue", color = "black") +
    theme_minimal() +
    labs(x = "Weighted edge betweenness", y = "Count")


```


##Arrange data 
```{r}

#[MA]make new column
Nodes_AT_root$Sub_community <- "AT_root"
Nodes_AT_soil$Sub_community <- "AT_soil"
Nodes_BO_root$Sub_community <- "BO_root"
Nodes_BO_soil$Sub_community <- "BO_soil"

#[MA]make a single df 
all_nodes <- bind_rows(
    Nodes_AT_root,
    Nodes_AT_soil,
    Nodes_BO_root,
    Nodes_BO_soil)

#[MA]Clasificación (Keystone/Hubs/Connectors)
all_nodes <- all_nodes %>%
    mutate(module_hub = ifelse(Zi > 2.5, "Module Hub", "Non-Hub"),
           module_connector = ifelse(Pi > 0.62, "Module Connector", "Non-Connector"),
           Keystone_Type = case_when(
               keystone_taxa == "Keystone" ~ "Keystone (manual)",
               #module_hub == "Module Hub" & module_connector == "Module Connector" ~ "Hub+Connector",
               module_hub == "Module Hub" ~ "Hub",
               module_connector == "Module Connector" ~ "Connector",
               TRUE ~ "Other"))

#[MA]clean taxonomy
all_nodes <- all_nodes %>%
    mutate(Domain = gsub("^d__", "", Domain),
        Phylum = gsub("^p__", "", Phylum),
        Class = gsub("^c__", "", Class),
        Order = gsub("^o__", "", Order),
        Family = gsub("^f__", "", Family),
        Genus = gsub("^g__", "", Genus),
        Species = gsub("^s__", "", Species))

#[MA]important families we've been highlighting 
main_families <- c(
  "Burkholderiaceae", 
  "Chitinophagaceae",
  "Comamonadaceae",
  "Flavobacteriaceae",
  "Oxalobacteraceae",
  "Rhizobiaceae",
  "Sphingobacteriaceae",
  "Streptomycetaceae",
  "Xanthobacteraceae") 

#[MA]make new column with these families written and otherwise, 'Other' will appear
all_nodes  <- all_nodes  %>%
    mutate(Family_simplified = ifelse(Family %in% main_families, Family, "Other"))

#[MA]check
head(as.data.frame(all_nodes)) #ok

#[MA]re-order for the plot
all_nodes <- all_nodes %>%
    mutate(Sub_community = factor(Sub_community, levels = c("AT_root", "BO_root", "AT_soil", "BO_soil")))


```

#Zi-Pi plot[Fig.4b]
##Color management
```{r}

#[MA] load full taxonomy table df
taxonomy_table_full <- readRDS("./R output/10_summary_results/taxonomy_table_full.rds")

main_families <- c("Burkholderiaceae", 
  "Chitinophagaceae",
  "Comamonadaceae",
  "Flavobacteriaceae",
  "Oxalobacteraceae",
  "Rhizobiaceae",
  "Sphingobacteriaceae",
  "Streptomycetaceae",
  "Xanthobacteraceae") 

#[MA]lets check to which order they belong to
main_families_order_class <- taxonomy_table_full %>%
    filter(Family %in% main_families) %>%
    dplyr::select(Family, Order, Class) %>%
    distinct()  # Si hay duplicados, los elimina

rownames(main_families_order_class) <- NULL #delete rownames 

#[MA]Get main orders 
main_orders <- unique(main_families_order_class$Order)

#[MA]let's just write them here so we can view them
main_orders <- c("Streptomycetales",
                 "Flavobacteriales",
                 "Burkholderiales",
                 "Rhizobiales",
                 "Chitinophagales",
                 "Sphingobacteriales")


#[MA]color vectors 
#[MA]Order colors - you can adapt these if needed
Order_col <- c(
    "Streptomycetales" = "#CC79A7",       # Morado (antes Rhizobiales)
    "Flavobacteriales" = "#0072B2",       # Azul oscuro
    "Burkholderiales" = "#FB8072",        # Rosa salmón fuerte
    "Rhizobiales" = "#F5C710",            # Amarillo fuerte
    "Chitinophagales" = "#56B4E9",        # Azul cielo
    "Sphingobacteriales" = "#009E73")      # Verde fuerte

#[MA] Family colors matching the new logic
Family_col <- c(
    "Burkholderiaceae" = "#F7A7A4",      # Burkholderiales
    "Chitinophagaceae" = "#56B4E9",      # Chitinophagales
    "Comamonadaceae" = "#F26B6C",        # Burkholderiales - tono diferente
    "Flavobacteriaceae" = "#0072B2",     # Flavobacteriales
    "Oxalobacteraceae" = "#FEE2C7",      # Burkholderiales - otro tono diferente
    "Rhizobiaceae" = "#FED700",          # Rhizobiales - amarillo más suave
    "Sphingobacteriaceae" = "#009E73",   # Sphingobacteriales
    "Streptomycetaceae" = "#CC79A7",     # Streptomycetales
    "Xanthobacteraceae" = "#D9A417")      # Rhizobiales - amarillo fuerte


#[MA]Add gray for others
order_colors <- c(Order_col, "Other" = "grey70")
family_colors <- c(Family_col , "Other" = "grey70")

```

[MA] Settings for plots 
```{r}
#Set plots theme
ggplot2::theme_set(theme_bw())

# Set default axis looks
axis_looks <- theme(axis.text.x = element_text(colour = "black", size = 9,
                                               face = "bold", angle=0, hjust=0.5))+ #text in X
  theme(axis.text.y = element_text(colour = "black", size = 8, face = "bold"))+ #text in y
  theme(axis.title=element_text(size=11, face = "bold"))+ #all texts
  theme(axis.title.y = element_text(margin = ggplot2::margin(t = 0, r = 10, b = 0, l = 0)))+
  theme(panel.grid.major = element_blank(), #no grid
        panel.grid.minor = element_blank()) #no grid
```

## Full plot[Fig.4b]
```{r}

#[MA]Full; also with regular nodes 

#[MA] Define keystone shapes and regular shape for Other
keystone_shapes <- c(
    "Keystone (manual)" = 23,
    "Hub" = 21,
    "Connector" = 22,
    "Other" = 16)  # Regular point

#[MA]plot it 
ggplot(all_nodes, aes(x = Zi, y = Pi,
                      shape = Keystone_Type,
                      fill = Family_simplified, color = Family_simplified)) +
    geom_point(data = all_nodes %>%
               filter(Keystone_Type == "Other"),  # Faded regular nodes (no outline)
               aes(color = Family_simplified), 
               size = 2, alpha = 0.5) + 
    geom_point(data = all_nodes %>%
                 filter(Keystone_Type != "Other"), # Important-Keystone with outline
               aes(fill = Family_simplified), 
               size = 4, color = "black", stroke = 0.3, alpha = 0.9) +  
    geom_hline(yintercept = 0.62 , linetype = "dotted", color = "gray",linewidth=0.5) +
    geom_vline(xintercept = 2.5, linetype = "dotted", color = "gray",,linewidth=0.5) +
    labs(x = "Within-module connectivity (Zi)",
         y = "Among-module connectivity (Pi)",
         title = "Keystone, Hubs & Connectors across Sub-communities",
         fill = "Family", color = "Family") +
    scale_shape_manual(values = keystone_shapes) +
    scale_fill_manual(values = family_colors,
                      guide = guide_legend(override.aes = list(shape = 21, size = 4))) +
    scale_color_manual(values = family_colors,
                       guide = guide_legend(override.aes = list(shape = 16, size = 4))) +
    facet_grid(~Sub_community) +
    axis_looks +
    theme(legend.position = "top")

```

[MA]Highlighting Mucilaginibacter
```{r}
#[MA]Now if you want to highlight Mucilinagibacter;
mucilagi_asvs <- all_nodes %>%
    filter(Genus == "Mucilaginibacter" & Keystone_Type %in% c("Connector", "Hub", "Keystone (manual)"))
print(mucilagi_asvs$Row.names)

#[MA]plot it 
ggplot(all_nodes, aes(x = Zi, y = Pi,
                      shape = Keystone_Type, fill = Family_simplified, color = Family_simplified)) +
    geom_point(data = all_nodes %>% filter(Keystone_Type == "Other"),
               aes(color = Family_simplified), 
               size = 1, alpha = 0.4) +  # Faded regular nodes (no outline)
    geom_point(data = all_nodes %>% filter(Keystone_Type != "Other"),
               aes(fill = Family_simplified), 
               size = 3, color = "black", stroke = 0.3, alpha = 0.9) +  # Important-Keystone with outline
    geom_text(data = mucilagi_asvs, aes(label = Row.names),
              size = 3, hjust = -0.2, vjust = -0.2, color = "black")+
    geom_hline(yintercept = 0.62, linetype = "dotted", color = "gray",linewidth=0.5) +
    geom_vline(xintercept = 2.5, linetype = "dotted", color = "gray",,linewidth=0.5) +
    labs(x = "Within-module connectivity (Zi)",
         y = "Among-module connectivity (Pi)",
         title = "Keystone, Hubs & Connectors across Sub-communities",
         fill = "Family", color = "Family") +
    scale_shape_manual(values = keystone_shapes) +
    scale_fill_manual(values = family_colors,
                      guide = guide_legend(override.aes = list(shape = 21, size = 4))) +
    scale_color_manual(values = family_colors,
                       guide = guide_legend(override.aes = list(shape = 16, size = 4))) +
    facet_grid(~Sub_community) +
    axis_looks +
    theme(legend.position = "top")

#[MA]Save plot
mapply(function(x)
  ggsave(
  plot = last_plot(),
  filename =x,
  path = "./Plots/06_network_analysis",
  scale = 1.5,
  width = 10.5,
  height = 6,
  units = "cm",
  dpi = 300),
x=c("Zi_Pi.svg","Zi_Pi.png"))


```

Only central taxa
```{r}
#[MA]plot only the central taxa
# Nos quedamos solo con los nodos importantes (descartamos "Other")
keystone_nodes <- all_nodes %>%
    filter(Keystone_Type != "Other")

keystone_shapes_K <- c(
    "Keystone (manual)" = 24,
    "Hub" = 21,
    "Connector" = 22)

#plot it 
ggplot(keystone_nodes, aes(x = Zi, y = Pi,
                      shape = Keystone_Type, fill = Family_simplified, color = Family_simplified)) +
    geom_point(data = all_nodes %>% filter(Keystone_Type == "Other"),
               aes(color = Family_simplified), 
               size = 2, alpha = 0.5) +  # Faded regular nodes (no outline)
    geom_point(data = all_nodes %>% filter(Keystone_Type != "Other"),
               aes(fill = Family_simplified), 
               size = 4, color = "black", stroke = 0.3, alpha = 0.9) +  # Important-Keystone with outline
    geom_text(data = mucilagi_asvs, aes(label = Row.names),
              size = 3, hjust = -0.2, vjust = -0.2, color = "black")+
    geom_hline(yintercept = 0.62, linetype = "dashed", color = "red") +
    geom_vline(xintercept = 2.5, linetype = "dashed", color = "red") +
    labs(x = "Within-module connectivity (Zi)",
         y = "Among-module connectivity (Pi)",
         title = "Keystone, Hubs & Connectors across Sub-communities",
         fill = "Family", color = "Family") +
    scale_shape_manual(values = keystone_shapes_K) +
    scale_fill_manual(values = family_colors,
                      guide = guide_legend(override.aes = list(shape = 21, size = 4))) +
    scale_color_manual(values = family_colors,
                       guide = guide_legend(override.aes = list(shape = 16, size = 4))) +
    facet_grid(~Sub_community) +
    axis_looks +
    theme(legend.position = "right")


```

#Venn Diagrams[Fig.4c] 
```{r}

keystone_nodes #In here you have only central taxa, let's split it by Plant Species and Compartment 

#[MA]: Make a list per Plant Sp and Rc
central_list <- split(keystone_nodes, keystone_nodes$Sub_community)

#Let's keep only the ASV names 
central_asv_list <- lapply(central_list, function(df) df$Row.names)

#[MA]Other option: let's load the list we already have with the JA-responsive ASVs per method
#important_ASV_list <- readRDS("./R output/10_summary_results/important_ASV_list.rds")

#[MA]keep only those of Network Analysis
#na_asv_list <- important_ASV_list$NA_ASV_list

#[MA]how many DA ASVs were per plant species and root compartment?
sapply(central_asv_list, length)

#[MA]load nVennR to make ven diagram
library(nVennR)

#[MA]Draw the Venn diagram
venn <- plotVenn(central_asv_list, systemShow = TRUE) #this will give you an .svg file that you can save 

```

#Save data
```{r}

#First let's check that we've saved what we need
central_asv_list

#[MA]Let's save the vector with ASV names 
saveRDS(central_asv_list, "./R output/06_network_analysis/NA_ASVs.rds")

#[MA] Let's save the environment 
save.image("./R output/env.6_network_analysis.RData")

#[MA] Now let's clean what we don't need for next script 7-Random Forest
rm(list = setdiff(ls(), c("normalization_listed")))

```

After this, plot network or load network file MeJA_pilot_networks.cys on cytoscape.  
Network analysis complete! proceed to script 7_random_forest
